cmd/atlas: move migrate command logging to cmdlog
Will add there the logging for the schema command
